reverted sum methor - it should actually return int. Add some optimization and tests for large scale + benchmark with actual stream API implementation

+++ src/main/java/ua/procamp/streams/stream/IntStream.java
@@ -22,7 +22,7 @@ public interface IntStream {
 
     int reduce(int identity, IntBinaryOperator op);
 
-    long sum();
+    int sum();
 
     int[] toArray();
 }
